app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium>=1.1.1",
    "mysql-connector-python>=9.3.0",
    "python-dotenv>=1.1.1",
    "stable-baselines3[extra]>=2.6.0",
    "pyyaml>=6.0",
    "tensorboard>=2.14.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "pandas>=2.0.0",
]
